moving files to lib
