+++ lib/dom.js
