+++ lib/entities.js
