+++ lib/sax.js
