+++ test/parse/test-define.js
@@ -1,3 +1,3 @@
 console.log(__dirname)
-exports.XMLSerializer = require('../../dom').XMLSerializer ;
-exports.DOMParser = require('../../dom-parser').DOMParser;
+exports.XMLSerializer = require('../../lib/dom').XMLSerializer ;
+exports.DOMParser = require('../../lib/dom-parser').DOMParser;
